% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.r
\name{compute_hidden_interest}
\alias{compute_hidden_interest}
\title{Compute the hidden interest}
\usage{
compute_hidden_interest(net_payment, govt_help)
}
\arguments{
\item{net_payment}{net payment}

\item{govt_help}{percentage of subsidzation the government will provide}
}
\description{
Sometimes loans will hide the interest if it causes the balance to exceed
the original principal. This interest may be capitalized if you change plans
or consolidate loans.
}
