% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.r
\name{compute_net_payment}
\alias{compute_net_payment}
\title{Compute the net yearly payment}
\usage{
compute_net_payment(balance, rate, ppm)
}
\arguments{
\item{balance}{current balance}

\item{rate}{interest rate}

\item{ppm}{payment per month}
}
\description{
Subracts student payment from accrued interest
}
\keyword{internal}
