% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.r
\name{compute_payments}
\alias{compute_payments}
\title{Compute payment plan}
\usage{
compute_payments(balance, rate, ppm, years, plan = c("repaye", "ibr"),
  subsidized = FALSE, ppm_increase = 0)
}
\arguments{
\item{balance}{current balance or principal}

\item{rate}{interest rate}

\item{ppm}{student payment per month}

\item{years}{number of years to calculate. Most debt forgiven after 25 years.}

\item{plan}{which payment plan to use}

\item{subsidized}{whether the government will help subsidze accrued interest}

\item{ppm_increase}{how much to increase the monthly payment amount per year. This is done with the hopes that a student will get a good job and be able to pay more over time.}
}
\description{
Compute payment plan
}
